# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage Customer API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/customer-store": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "description": "Create a new customer",
                "operationId": "f44a116b72c6f501c5290e8ee7f362a5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jose",
                                        "nullable": "false"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "jose@gmail.com",
                                        "nullable": "false"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "999999999",
                                        "nullable": "false"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1999-01-01",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer registered successfully"
                                        },
                                        "customerId": {
                                            "type": "integer",
                                            "example": "123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field must be a valid email address."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The email field must be a valid email address."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to register the customer"
                                        },
                                        "customerId": {
                                            "type": "integer",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer-update": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "description": "Update a customer",
                "operationId": "6170da934c16f7d3c13bc20a073e03b6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "10",
                                        "nullable": "false"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jose",
                                        "nullable": "false"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "jose@gmail.com",
                                        "nullable": "false"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "999999999",
                                        "nullable": "false"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1999-01-01",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer registered successfully"
                                        },
                                        "fields": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jose"
                                                },
                                                "email": {
                                                    "type": "email",
                                                    "example": "jose@gmail.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "999999999"
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-01-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The selected id is invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The selected id is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update the customer"
                                        },
                                        "fields": {
                                            "type": "object",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer-delete": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "description": "Delete a customer",
                "operationId": "d148122cb85cb31fa66f7a91d65205e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "123",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer deleted successfully"
                                        },
                                        "wasDeleted": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The selected id is invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The selected id is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete the customer"
                                        },
                                        "wasDeleted": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}